Add files via upload
This file is the re-creation of the VoiceForge demo webpage. Now you can type as much characters as you need, the only problem is  audio playback when playing what you wrote down.
Morning! I do not own this code, all credit goes to the creators of VoiceForge. I only did this so I can play around with the code and see what things I can change.