Quick and dirty monkeypatch for the VoiceForge demo page requesting an HTTP version of JQuery.
A patch that is for the VoiceForge demo page requesting an HTTP version of JQuery. It may have been the reason why VoiceForge would not play audio for some people if they haven't turned on allow unsecured content. This problem started in summer of 2018 ware the text you entered was not playing.